feat: id type fixes

+++ internal/domain/checkin.go
@@ -12,9 +12,9 @@ const (
 )
 
 type Checkin struct {
-	ID int `json:"id"`
+	ID int64 `json:"id"`
 
-	GoalID int `json:"goal_id"`
+	GoalID int64 `json:"goal_id"`
 
 	CheckinDate time.Time     `json:"checkin_date"` // YYYY-MM-DD
 	Status      CheckinStatus `json:"status"`
